Work on CLI token refresh
